Started migration of Reverse Linked List
